projects: ad353xr: Add minor updates
Update extsetting with active platform to 2 for stm32.
Update code to fix complie time warnings.

+++ projects/ad3530r_iio/app/app_config_stm32.c
@@ -351,7 +351,7 @@ void receivecomplete_callback(DMA_HandleTypeDef* hdma)
 			*  after the first linear/normal dma transfer is completed */
 			DMA2_Stream3->CR &= ~DMA_SxCR_EN; // Disable DMA
 			DMA2_Stream3->CR |= DMA_SxCR_CIRC; //circular mode enable
-			DMA2_Stream3->M0AR = global_iio_buff + 2;	// update memory address
+			DMA2_Stream3->M0AR = (uint32_t)(global_iio_buff + 2);	// update memory address
 			DMA2_Stream3->NDTR = num_of_samples * 2; // update NDTR
 
 			DMA2_Stream3->CR |= DMA_SxCR_EN; // Enable DMA
